Rename piecesService to pieceAccessLevelService & add docstring
